docs/source/Compute/Lambda-Root/Deploy-Lambda-with-AWS-CDK/requirements.txt
aws-cdk-lib==2.111.0
constructs>=10.0.0,<11.0.0
boto_session_manager>=1.7.1,<2.0.0
